% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FKF-package.R
\docType{package}
\name{FKF-package}
\alias{FKF}
\alias{FKF-package}
\title{FKF: Fast Kalman Filter}
\description{
This is a fast and flexible implementation of the Kalman filter and smoother, which can deal with NAs. It is entirely written in C and relies fully on linear algebra subroutines contained in BLAS and LAPACK. Due to the speed of the filter, the fitting of high-dimensional linear state space models to large datasets becomes possible. This package also contains a plot function for the visualization of the state vector and graphical diagnostics of the residuals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://waternumbers.github.io/FKF/}
  \item \url{https://github.com/waternumbers/FKF}
  \item Report bugs at \url{https://github.com/waternumbers/FKF/issues}
}

}
\author{
\strong{Maintainer}: Paul Smith \email{paul@waternumbers.co.uk} (\href{https://orcid.org/0000-0002-0034-3412}{ORCID})

Authors:
\itemize{
  \item David Luethi
  \item Philipp Erb
  \item Simon Otziger
  \item Daniel McDonald \email{daniel@stat.ubc.ca}
}

}
\keyword{internal}
